refactor(frontend): resolve pr comment bb-204
