+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/enums/chart-slice-color.enum.ts
