+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/enums/enums.ts
@@ -1,2 +1,2 @@
 export { BalanceWheelChartDataConfig } from "./balance-wheel-data-config.enum.js";
-export { ChartSliceColor } from "./chart-slice-colors.enum.js";
+export { ChartSliceColor } from "./chart-slice-color.enum.js";
