+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/helpers/generate-gradient-color/libs/maps/maps.ts
@@ -1,0 +1,1 @@
+export { labelToBackgroundColor } from "./label-to-background-color.map.js";
