+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/helpers/helpers.ts
@@ -1,2 +1,2 @@
-export { generateGradientColor } from "./generate-gradient.helper.js";
+export { generateGradientColor } from "./generate-gradient-color/generate-gradient-color.js";
 export { generateRandomData } from "./generate-random-data.helper.js";
